graph_classifier/requirements.txt
torch==2.0.1
torchaudio
torchvision
torchmetrics==0.11.4
einops
mne==1.6.0
hydra-core==1.3.2
wandb
termcolor
pymatreader